Move 'read' and 'write' to session.Session
